refactor: promote baseline virtual-package helpers onto ResolvedEnvironment
Both `_baseline_virtual_package_env` and `_apply_env_overrides` in
`lockfile.py` only existed to scope cross-platform solver env vars
derived from a resolved environment's `system_requirements`. Move them
onto `ResolvedEnvironment` so the data and the derived behaviour live
together:

- `ResolvedEnvironment.baseline_virtual_package_env(platform)` — pure
  method returning the `CONDA_OVERRIDE_*` dict tests assert on.
- `ResolvedEnvironment.scoped_solver_env(platform)` — contextmanager
  method that `_solve_for_records` uses to scope those overrides around
  a solver call.

Drops the two module-level private helpers from `lockfile.py` and the
associated `os` / `contextlib` imports there. Tests and DESIGN.md are
updated to the new API.
